[Components/Tests]: Split components tests into multiple files. (#881)
